Fixes some types on OpenBSD

+++ src/file/stat.cr
@@ -65,9 +65,9 @@ lib LibC
   elsif openbsd
     struct Stat
       st_dev : Int32
-      st_ino : Int32
+      st_ino : UInt64
       st_mode : LibC::ModeT
-      st_nlink : UInt16
+      st_nlink : UInt32
       st_uid : UInt32
       st_gid : UInt32
       st_rdev : Int32
